backendMain/requirements.txt
djangorestframework-jwt
Django==2.1
djangorestframework==3.8.2
djangorestframework-simplejwt==3.2.3
pkg-resources==0.0.0
PyJWT==1.6.4
pytz==2018.5
